[CP] Introduce flex_attention_wrapper
**Motivation**

Since FlexAttention and SDPA are both functions, not modules, we have tried numerous mechanisms to dispatch FlexAttention and SDPA to customized call paths so that we can inject the CP logic. Unfortunately, all of these approaches have their own composability issues with different techniques.

**Candidate Approaches**

1. Ask users to write a module to wrap FlexAttention/SDPA and use `parallelize_module` to install a forward hook.

   - Pros: This is similar to how we do TP.
   - Cons: 1) It is cumbersome for users as they need to create a new module. 2) We need two places to parallelize the CP, as a context_parallel context switch is still required for splitting the inputs.

~~2. Provide a function wrapper.~~

~~   - Pros: Users just need to replace their FlexAttention/SDPA calls with the wrapper.~~
~~   - Cons: It is not the same API, though we can maintain the API signatures to be the same as the core API.~~

**Summary**

~~This PR implements approach 2 but uses a nn.Module to mimic a function wrapper so that we can record CP state in the module instead of leaking it to the `_attention` module.~~

We changed this PR to implement option 1 as people like option 1 due to the consistency with the existing parallelisms. But this PR can also serve the foundation to implement option 2, which was the early version of this PR.

This PR also removes the unit test of TorchFunction dispatching method as it has an unsolveable composability issue with SAC. The TorchFunction dispatching code will be deleted from CP code base once we verify the E2E with the module hook.

ghstack-source-id: 59607823cd0451f5ba6efe9369230d9385029bf9
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/162542